Make explicit uint16_t call
Otherwise this example code fails to compile for platforms where the
default "int" isn't a 16 bit int

+++ examples/readTest/readTest.ino
@@ -69,7 +69,7 @@ void setup() {
   Serial.println(F("Looking for testStr..."));
   
   char testStr[15];
-  unsigned int filePtr, fileSize;
+  uint16_t filePtr, fileSize;
   r = OSFS::getFileInfo("testStr", filePtr, fileSize);
 
   if (r == notfound)
